Updated test config and fixed tests

+++ metrics/processors/namespace_aggregator_test.go
@@ -76,9 +76,9 @@ func TestNamespaceAggregate(t *testing.T) {
 
 	m1, found := namespace.MetricValues["m1"]
 	assert.True(t, found)
-	assert.Equal(t, 110, m1.IntValue)
+	assert.Equal(t, int64(110), m1.IntValue)
 
 	m3, found := namespace.MetricValues["m3"]
 	assert.True(t, found)
-	assert.Equal(t, 30, m3.IntValue)
+	assert.Equal(t, int64(30), m3.IntValue)
 }
